backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
google-generativeai==0.3.2
faiss-cpu==1.7.4
langchain==0.1.4
langchain-google-genai==0.0.6
pypdf2==3.0.1
xgboost==2.0.3
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3
pydantic==2.5.3
python-multipart==0.0.6
jinja2==3.1.3
reportlab==4.0.8
aiofiles==23.2.1
websockets==12.0
